Update encoder to encode GOP length == frame rate, and update tests

+++ core/video/include/VideoDecoderSession.h
@@ -113,7 +113,7 @@ protected:
                                     assert(i);
                                     break;
                                 } else {
-                                    while (decoder.frameNumberQueue().write_available() < numberOfFrames)
+                                    while (decoder.frameNumberQueue().write_available() < (long unsigned int)numberOfFrames)
                                         std::this_thread::sleep_for(std::chrono::milliseconds(10));
 //                                        std::this_thread::yield();
                                 }
